minimum te/tr with double

+++ src/seq/flash.c
@@ -229,12 +229,12 @@ static double gradient_time_after_RO(const struct seq_config* seq)
 	return 0.;
 }
 
-long min_tr_flash(const struct seq_config* seq)
+double min_tr_flash(const struct seq_config* seq)
 {
-	long time_ro_rf = MAX((seq->sys.min_duration_ro_rf - (long)(seq->sys.grad.max_amplitude * seq->sys.grad.inv_slew_rate)), seq->sys.coil_control_lead);
+	double time_ro_rf = MAX((seq->sys.min_duration_ro_rf - seq->sys.grad.max_amplitude * seq->sys.grad.inv_slew_rate), seq->sys.coil_control_lead);
 	double time_gradients = gradient_time_after_RO(seq);
 
-	long time_after_RO = MAX((double)time_ro_rf, time_gradients);
+	double time_after_RO = MAX(time_ro_rf, time_gradients);
 
 
 	struct grad_trapezoid last_ro;
@@ -247,7 +247,7 @@ long min_tr_flash(const struct seq_config* seq)
 }
 
 
-void min_te_flash(const struct seq_config* seq, long* min_te, long* fil_te)
+void min_te_flash(const struct seq_config* seq, double* min_te, double* fil_te)
 {
 	double ro_deph_time = available_time_RF_SLI(1, seq);
 	double inter_duration_READ = MAX(ro_deph_time, seq->sys.grad.max_amplitude * seq->sys.grad.inv_slew_rate);
